# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a voda api server.",
        "title": "Voice Of Diary API (voda)",
        "contact": {
            "name": "API Support",
            "url": "https://minkj1992.github.io",
            "email": "minkj1992@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "exchange-diary-b4mzhzbzcq-du.a.run.app",
    "basePath": "/v1",
    "paths": {
...
        "/rooms": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "참여중인 교환일기방 리스트",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "List rooms",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint",
                        "description": "page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "uint",
                        "description": "page offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.listResponseRoom"
                        }
                    },
                    "400": {
                        "description": ""
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "교환일기방 생성",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "create a room",
                "parameters": [
                    {
                        "description": "교환일기방 생성요청 body",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.postRequestRoom"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.postResponseRoom"
                        }
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        },
        "/rooms/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "교환일기방 상세",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "get a room",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint",
                        "description": "교환일기방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.detailResponseRoom"
                        }
                    },
                    "400": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "교환일기방 삭제",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "(debug only) delete a room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "교환일기방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "교환일기방 업데이트 (master only)\n1. 작성주기 변경 (period)\n2. 코드/힌트 변경 (code, hint)\n3. 작성순서 변경(orders)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "update a room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "교환일기방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "교환일기방 수정 요청 body",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.patchRequestRoom"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.patchResponseRoom"
                        }
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        },
        "/rooms/{id}/join": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "교환일기방 참여코드 체크 후, 교환일기방 멤버로 추가",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "join a room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "교환일기방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "교환일기방 참여 요청 body",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.verifyRequestRoom"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/rooms/{id}/leave": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "교환일기방 나가기\n1. 교환일기방 마스터일 경우\n2. 교환일기방 멤버일 경우",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "leave a room",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "교환일기방 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.detailResponseRoom": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "hint": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isMaster": {
                    "type": "boolean"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.responseMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "period": {
                    "type": "integer"
                },
                "theme": {
                    "type": "string"
                },
                "turnAccountId": {
                    "type": "integer"
                },
                "turnAccountName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "controller.listResponseRoom": {
            "type": "object",
            "properties": {
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.responseRoom"
                    }
                }
            }
        },
...
        "controller.patchRequestRoom": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "hint": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "period": {
                    "type": "integer"
                }
            }
        },
        "controller.patchResponseRoom": {
            "type": "object",
            "properties": {
                "roomId": {
                    "type": "integer"
                }
            }
        },
        "controller.postRequestRoom": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "제민욱"
                },
                "hint": {
                    "type": "string",
                    "example": "레오의 본명은?"
                },
                "name": {
                    "type": "string",
                    "example": "고영희방"
                },
                "period": {
                    "type": "integer",
                    "example": 5
                },
                "theme": {
                    "type": "string",
                    "example": "1"
                }
            }
        },
        "controller.postResponseRoom": {
            "type": "object",
            "properties": {
                "roomId": {
                    "type": "integer"
                }
            }
        },
        "controller.responseMember": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "profileUrl": {
                    "type": "string"
                }
            }
        },
        "controller.responseRoom": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.responseMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "controller.verifyRequestRoom": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}